option[l]の出力要素をLdataクラスを作ってそちらに移管
